+++ PeasyMotion/PeasyMotionPackage1.cs
@@ -35,6 +35,7 @@ namespace PeasyMotion
         public const int PeasyMotionLineJumpToWordEndingId = 0x0103;
         public const int PeasyMotionJumpToDocumentTab = 0x0104;
         public const int PeasyMotionJumpToLineBegining = 0x0105;
+        public const int PeasyMotionTwoCharJump = 0x0106;
         public const int bmpHighlightText = 0x0001;
     }
 }